Adding UWP Capability support to Platform Editor
Adding UWP Capability and TargetDeviceFamily support to the 'Platforms -
UWP' page in Editor.

This check-in allows for selecting the TargetDeviceFamily (Universal,
Holographic, Desktop, etc.) and enabling/disabling Capabilities from
within the Editor. Supported Capability types include: uap2:Capability,
uap:Capability, DeviceCapability, and Capability. When a capability is
toggled within the Editor, they are added/removed from a TARRAY that
matches their type. These capability arrays are saved in the Game's
DefaultEngine.ini and referenced by the UWPEngine.ini, which is used
during packaging. Special handling is required for the uap2:Capability
and DeviceCapability types in order to pass AppxManifest validation.
Adding capabilities to the AppxManifest file is handled within
ManifestGenerator.cs.

This change also includes fixes for broken image links and background
color options on the Platforms - UWP settings page.