The DXGI factory pointer stored by the D3D11 RHI has to be the same one that was used to create the adapter.
The code was creating one DXGI factory when it made the device, but then it made a second factory which was stored in FD3D11DynamicRHI and used for subsequent calls. This resulted in a subtle problem with swapchain creation: CreateSwapChain was called on the second factory pointer, but internally it doesn't care and uses the parent factory of the device, i.e. the first factory we were making. When we called MakeWindowAssociation on the same (second) pointer, it didn't work, because it was not the factory which really owned the swapchain. This meant that DXGI was still hooking our window proc and catching alt+enter and tried to handle fullscreen for us, which broke all sorts of things.

#jira FORT-411923, UE-122160
#rnx
#lockdown Michal.Valient
#rb Christopher.Waters

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 17642278 in //UE4/Release-4.27/...
#ROBOMERGE-BOT: RELEASE (Release-4.27 -> Release-4.27-Chaos) (v874-17637634)

[CL 17642370 by mihnea balta in 4.27-Chaos branch]