Fix crash when streaming in sublevels containing lights.
The denoiser code was queueing up RDG texture extraction into memory owned by a TMap which can be resized during the frame. This is similar to the problem with bloom history targets fixed in CL 17606212, but in this case we can't use a TUniquePtr to manage the memory, because FPreviousViewInfo is copied around by the renderer (e.g. in FSceneRenderer::PreVisibilityFrameSetup). The simplest solution is to use a shared pointer.

#jira UE-115699
#rb guillaume.abadie
#lockdown michal.valient
#rnx

#ROBOMERGE-AUTHOR: mihnea.balta
#ROBOMERGE-SOURCE: CL 17796602 in //UE4/Release-4.27/...
#ROBOMERGE-BOT: RELEASE (Release-4.27 -> Release-4.27-Chaos) (v883-17842818)

[CL 17843910 by mihnea balta in 4.27-Chaos branch]