Fixed Blueprint type autoconversion for variable nodes. (resubmit)
Unlike other nodes, the variable node is the only that has a ArePinTypesCompatible check. This created a discrepancy: type changes on function nodes would preserve the link and create a compilation error, but type changes on variable nodes would orphan the pin.

For consistency, this removes the ArePinTypesCompatible check. More importantly, this also fixes auto type conversion, which will analyze the link and insert a cast node at a later stage. However, there's also some legacy cruft that only applied to object pins with mismatched types. This particular edge case has been preserved, but it might need follow up to see if it's still needed.

Additionally, fixed a logic bug with bCanMatchSelfs.

#rb phillip.kavan

[CL 26032689 by dave jones2 in ue5-main branch]