Add support for XR_FB_foveation OpenXR extension.
Added a foveation swapchain in OpenXRHMD and implemented IVariableRateShadingImageGenerator to update the foveation swapchain to the latest settings and provide the foveation images to RenderGraph. Also added the possibility in OpenXRHMD_Swapchain to create a foveation swapchain instead of a normal swapchain. This only works when using either Vulkan or OpenGL ES.

#jira UE-184914
#rb christopher.fiala robert.srinivasiah

[CL 26075136 by erica stella in ue5-main branch]