Configure FakeStereoRenderingDevice based on viewport settings.
This changes the default FOV of the emulated stereo view to 90, which matches the default FOV for single view. The width and height of the FakeStereoRenderingDevice, which are used to compute the projection matrix, are now also initialized based on the viewport resolution (instead of default constants of 640 and 480). The main purpose of this change is so that running a project with "-game -windowed -emulatestereo" with twice the resolution width of the non-stereo version will be roughly equivalent in terms of rendering.

#jira UE-164902
#rb robert.srinivasiah

[CL 26105088 by jeffrey liu in ue5-main branch]