Sequencer: Fix leaking pre animated state for tracks set to keep state. Looping sequences with spawnables in subscenes could cause memory to continually get allocated in pre animated storage for each time spawnables were created. This was occurring for sections that were set to 'Keep State'. Even though states were not being cached, memory was being allocated in the cache tied to the object keys because 'MakeEntry' was being called. This memory was not even required as we weren't ever caching into it or restoring state. To fix this, logic has been reorganized in the CachePreAnimatedState functions to check whether we are meant to be tracking state before creating entries.
In addition on garbage collection we go through the metadata and remove any metadata for objects that may have been garbage collected, e.g. Spawnables. This could also be a source of leaks for the looping sequence states.

 Also, removed the function 'ShouldSavePreAnimatedState' as it wasn't being used and simplified the logic.

Also fixed a case where Data Layer preanimated state was firing an ensure by filtering on the RestoreState tag, which should be unnecessary as the BeginTrackingEntity function already handles that case. So that's been removed.


#jira UE-188917
#jira UE-164431
[REVIEW] [at]ludovic.chabant [at]sebastien.lussier
[FYI] [at]andrew.rodham [at]max.chen

[CL 26143515 by david bromberg in ue5-main branch]