[Chaos VD] Making all delegates from ChaosVDPlaybackController trigger on the Game Thread.
During Trace analysis, we fill up the ChaosVDRecording structure. At the same time, as soon a new frame data is available, the playback controller will process it (allowing to play a recording before it finished loading).

That is useful to reduce waiting times, but is also a requirement for remote debugging using trace as in that mode we will be receiving frames constantly as we play.

There are a few issues with this and the current implementation:

- Trace Analysis updating the recording, and the playback controller processing it will trigger an UI update, and when that is done outside the game thread, it will crash.
- Playback is done in the GameThread, to do so we get a reference to the solver frame data from the recording we want to play. if the container where that data lives is re-allocated while Trace Analysis is adding data, the tool will crash.

This CL ensure all events reporting changes used by UI code are triggered in the Game Thread instead of the Trace Analysis thread.

In addition a new RW lock was added to the recording structure (plus AssumesLocked versions of existing methods), to protect the data from being accessed during playback when we are writing a new frame.

The Assumeslocked version of some methods are required because we need to lock the data between multiple scopes, and in some scenarios in a  recursive manner. An alternative would to use a recursive mutex.

#jira UE-187979

[CL 26241534 by sergio gardeazabal in ue5-main branch]