[Backout] - CL26284902 [FYI] michael.short Original CL Desc ----------------------------------------------------------------- [Backout] - CL26281988 [FYI] michael.short Original CL Desc ----------------------------------------------------------------- Unshelved from pending changelist '26281904':
Quartz Subscriber Optimization

During the creation of this class, it would subscribe itself to the Quartz Subsystem, so that it can receive regular tick updates. The unsubscribe was happening exclusively in the destructor. Therefore for managed objects, they weren't being unsubscribed until the object was fully deleted by the garbage collector.

I wrapped up the subscription code inside a new QuartzSubscriptionToken. This class adds a few more checks for sanity, and prevents things from being subscribed multiple times, or unsubscribed when its not even subscribed. The Actor/Component objects that derive from this class now unsubscribe in the EndPlay function call - they subscribe in BeginPlay. This prevents the constant build up in subscribed objects until the garbage collector runs.

[CL 26403602 by michael short in ue5-main branch]