Sequencer: patch up bound object pointers in the property instantiator system when an object gets replaced
Blueprint actors are recompiled very often, such as when they are moved. This caused pointers in the property instantiator to go stale, and the new object to now have any properties resolved. It was corrected on the next sequencer update, but it meant that keying the moved object would fail to recompose values.

#jira UE-189422
#rb david.bromberg

[CL 26412103 by ludovic chabant in ue5-main branch]