MoviePipeline: Added "Apply CVar Preset" node to the graph, and integrated cvars into the graph pipeline.
Because the graph pipeline pulls in cvars from multiple node types (currently the "Apply CVar Preset" and "Set CVar Value" nodes), an individual node should not be responsible for applying cvars. A new class that deals with applying cvars in the pipeline, FMovieGraphCVarManager, was created to deal with this.

The cvar nodes were also locked down to the Globals branch. In this version of the graph, cvars will only be applied once before a render to avoid issues with setting cvars to different values on various branches, or setting cvars in the middle of a render (thread synchronization problems).

#jira UE-189607
#rb matt.hoffman
#rnx

[CL 26412481 by chad williams in ue5-main branch]