Sequencer: fix camera cut flag not being set when spawnable cameras are the last in control
This was because calling GetViewTarget does some internal checks. If a spawnable camera has just been unspawned (i.e. the actor is PendingKill) then this trips those checks and the view target is reset to the pawn before we can even do anything. Because we are restoring the view target to the pawn, this is registered as "not changing anything" and we skip everything in UpdateCameraCut, including setting the camera cut flag.

We therefore don't use GetViewTarget, and grab the view target directly from the public member to manage things by hand.

#jira UE-190045
#rb max.chen

[CL 26550161 by ludovic chabant in ue5-main branch]