Sequencer: Fixed race condition caused by async children being able to signal the main thread after completion while a worker thread was finishing iterating its children
The race here stems from the fact that a child task signals its parent for completion if it was the last one. In doing so, it could also signal the main thread to finish waiting on tasks, if it was the last task to be executed. This is problematic because there is still a thread iterating the child tasks at the same time as the last one is completed, resulting in a race between the iteration finishing successfully, and the main thread exiting out of our loop and potentially reallocating the array.

The fix is to prevent children from completing their parents while the loop is running by adding an artificial count for the iteration. If there are no children outstanding after the loop, CompleteTask is called for the Parent.

#rb Max.Chen

[CL 26769248 by andrew rodham in ue5-main branch]