Extended TEDS' Compatibility layer with the ability to look up objects that are referenced but aren't automatically mapped to a row.
A use case for this change is to look up a the row a component is stored on. The owning actor is mapped to a row while the components on that actor are mapped to columns. This allows looking up the correct by getting the owning actor from the component and then look that up instead.

#jira UE-189623
#rb kurtis.schmidt

[CL 26776959 by ronald koppers in ue5-main branch]