Fix buoyancy volume double-counting.
This works by adding a sparsearray of bitarrays. The sparsearray is indexed by the particle unique indices. The bitarrays are indexed by shape indices. If a bit is set, then skip processing that shape. If not, then process a shape and set it.

This change also introduces a division between application of buoyant forces and dedection of submersions, in hopes that we can remove the manual integration bit at some point.

[CL 26793628 by steven barnett in ue5-main branch]