In FLandscapeBrushCircle, moved cleanup of stale weak object ptrs entries for both BrushMaterialInstanceMap and BrushMaterialComponents from AddReferencedObjects to Tick.
The issue was caused by the cleanup of BrushMaterialInstanceMap which could remove entries from the map if the value (i.e. the material instance) got stale, while keeping the corresponding entry in the BrushMaterialComponents (since the landscape component itself was still perfectly fine), which led to an assert/crash in Tick, when we were trying to remove entries from the map, assuming there was one for each component in BrushMaterialComponents.

Now we first cleanup BrushMaterialInstanceMap from stale keys (i.e. null landscape component weak object ptrs) and then also cleanup stale entries in BrushMaterialFreeInstances. BrushMaterialComponents doesn't need to be cleaned up because of the way the code already works, at the end of the Tick it only contains valid components.


#rb Roey.Borsteinas

[CL 26817714 by chris tchou in ue5-main branch]