Chaos Cloth: Caching Up/Down transitions for LODs on the terminal node.
Calculating the Up/Down transitions can be expensive, especially for heavier meshes. However, since it happens at the terminal node, they are recalculated anytime anything changes on the asset, including simulation parameters, which do not affect the transitions.

This adds the ability to cache the transitions on the terminal node and hash the inputs. The transitions are reused if the hash didn't change.

#rb kriss.gossart

[CL 26861548 by alex mcadams in ue5-main branch]