Refactored FMassEntityManager::GetMatchingArchetypes and renamed to GetMatchingArchetypes. The change extends what's possible to query - now a query declaring some tag requirements and only optional fragment requirements will still fetch some results (assuming there are archetypes matching). The change also made the new GetMatchingArchetypes use FMassFragmentRequirements as a parameter, rather than a FMassEntityQuery.
As part of the change FMassEntityManager::AllArchetypes array has been added that collects all archetypes created by the manager. This lets us find archetypes based on tags alone a bit faster. This addition is also a prerequisite for a future change to FMassArchetypeHandle to not store a shared pointer but rather an index to the FMassEntityManager owned array.

[CL 26948811 by mieszko zielinski in ue5-main branch]