Remove UE_TRACE_ENABLED check
It should be possible to launch UTS without having tracing enabled in the application. This is important to allow that for example UnrealInsights can launch the trace server. With this the log messages has to be emitted on the LogCore channel, as the LogTrace only exists when tracing is enabled.

#jira UE-190571
#rb ionut.matasaru, catalin.dragoui
#lockdown mark.lintott

[CL 27029990 by johan berg in ue5-main branch]