Debug Draw: fix usage of input scene color rect + output view rect
Previously, we relied on the fact that the target rect matched into the source color/depth textures.
However, with TSR + Stereo rendering, TSR outputs a monoscopic texture, while DebugPrimitives is
drawing into a double-wide stereo texture. Because of this, we need to generate UVs from the target rect,
and sample from the input texture rects.

It's unclear why this behavior was changed in 25024370. I think the original CL wanted to 'target' the target
view rect, but that shouldn't have affected the source rect.

#jira UE-191760
#rb Jon.Cain

[CL 27085390 by robert srinivasiah in ue5-main branch]