Verse heap medium pages should have out-of-line headers.
Libpas supports medium pages that have inline or out-of-line headers. Out-of-line headers are better. Prior to this change, the Verse heap used inline medium page headers, for silly complex reasons (a prior design of the Verse heap would have benefited from it). Now, it's the opposite, the Verse heap layout makes out-of-line medium page headers easy, since the chunk map can be used as the page header map. So, this change switches medium pages to use out-of-line headers via chunk map lookup.

I have a hunch this'll make SOL-4903 a little easier to do. And, it's worth doing this anyway, since the inline medium page headers are a holdover from an abandoned design.

#rb saam.barati

[CL 27096248 by phil pizlo in ue5-main branch]