Sequencer: Improved handling of garbage entities inside UMovieScenePropertyInstantiator::OnCleanTaggedGarbage
Previously it was possible for OnCleanTaggedGarbage to pick up and process new entities and create blend channels and outputs for them, resulting in an output entity with NeedsLink. This was problematic becuase if that blend output happened to relate to a group of entities that ends up being considered as a fast-path entity at the next instantiation phase (probably because a sequence was stopped, or a track ended), we would try and add NeedsUnlink to this blend output when it still has NeedsLink. This is not a supported combination.

The fix is to only process expired entities in OnCleanTaggedGarbage and keep around the invalidated property indices until the next full run. This prevents us from being able to create entities inside OnCleanTaggedGarbage.

Furthermore, we now iterate and detect the destruction of blend output entities and null out such references to ensure they don't get recycled.

[REVIEW] [at]max.chen, [at]Ludovic.Chabant
#rb Ludovic.Chabant

#localization none
#tests Ran all EngineTests, AdHoc testing on FrontEnd, AdHoc BR testing in PIE, Preflight w/CreativeTest,EdTest,UEFN

[CL 27146366 by andrew rodham in ue5-main branch]