First iteration of TEDS memento system
This CL introduces a mechanism to opt into creating mementos of columns for mementoized rows.
 Mementoization is the process of copying the a subset of a row's state and storing it to be later read to restore the state of a row.

The heart of this process is in the MementoTranslator.  These are CDOs which define the mapping between the column that will be mementoized and the memento itself.  They also define the mechanism to covert between columns and mementos.  This initial implementation only allows for column to memento operations, though extension to allow the reverse will be added later.

MementoTranslators have two levels to the API.
Level 1: The DefaultMementoTranslator is the easiest to set up.  It will generate a runtime struct of the mementoized column based on the UPROPERTIES of the column and has code that will automatically populate the memento from the column.  There is an example of it's use in UTypedElementLocalTransformColumnMementoProcessor.
MementoTranslator:
Level 2: The MementoTranslator can be overridden by user code to get more control over the mementoization process.  It allows explicitly defining the translation process. This can be useful for optimizing the size of the mementos themselves or the speed at which mementoization occurs.  The UTypedElementPackagePathMementoTranslator is an example of using this API.

Rows which want to opt into mementoization need to add a FTypedElementMementoOnDelete column and a placeholder unpopulated Memento row.  This MementoSystem will watch for rows deleted with this column and any mementoized column defined by the available MementoTranslators and invoke the appropriate mementoization operators.

Memento rows will exist during the lifetime of a mementoized row and will be destroyed automatically after 120 frames to avoid continued expansion of the memento tables.

By default, mementoization is disabled for objects in the Compatibility layer.

CVARS:
`teds.mementos.enable 1` - enable mementoization on newly created objects.
`teds.mements.keepframes N` - The number of frames that a populated memento sticks around after the mementoized row is deleted

#jira UE-192033
#rb ronald.koppers

[CL 27219901 by logan buchy in ue5-main branch]