[PCG] Push higen grid size resolve pass deeper, from schedule to compile.
This way higen considerations are made before caching and this removes the 'patching' that currently happens in schedule. A set of compiled tasks are cached per execution grid size.

A next step will be to cull tasks that are not relevant for a grid size, which will be the major benefit of caching per grid size. And static switches will hopefully also allow culling here too.

#rb julien.lheureux, ryan.buehler
#rnx
#jira UE-193241

[CL 27269622 by huw bowles in ue5-main branch]