Update how RemoveOptionPin works for Dataflow nodes. These are currently only used by a couple of cloth nodes.
The problem was that we weren't updating the FDataflowNode's connection data that the connection was broken. This would eventually cause a crash because the connection data was deleted, but there were still pointers hanging around with them.This was fixed by calling BreakAllPinLinks with bNotifyNodes = true inside UDataflowEdNode::RemoveOptionPin and then deleting the pin AFTER breaking the link. I did the latter by breaking the RemovePin method into two methods, a const method GetPinToRemove which just returns which pin (if any) to remove, and a non-const method OnPinRemoved which will clean up the DataflowNode's connection.

Note that I did not update the way AddOptionPin works because getting the data required to create the input is easiest in one call to the derived node, rather than splitting. Also there is no existing callback for OnPinAdded at the ed graph level.

#rb kriss.gossart cedric.caillaud

[CL 27281487 by Alex McAdams in ue5-main branch]