#jira: FORT-638338, FORT-645380
Sort the plugins before loading all built in plugins so the leaf plugins are loaded first. This gives a more accurate count to how long a plugin takes to load and removes the need to perform extra work subtracting to calculate the exclusive plugin load time. We are now always using the exclusive plugin load time by sorting.

[REVIEW] [at]Justiin.Marcus [at]Will.Brown [at]Bob.Tellez

[CL 27315862 by eric knapik in ue5-main branch]