[Backout] - CL27337684 [FYI] massimo.tristano Original CL Desc ----------------------------------------------------------------- Introduced Material Translation DDC caching.
This change implements a caching mechanism for material translation results. After translating a material, the intermediate results are now stored in the DDC (Derived Data Cache) for future retrieval. When the material translator is requested to translate a material, it first checks the DDC to see if th combination of specified base material, parameter sets, and platform-specific data is already present in the cache. If this key is found, the intermediate results are pulled and returned, avoiding expensive material translation and decoupling material translation from material shader generation.

The significant advantage of this change is that modifications to the material template source file or any dependent shader file no longer trigger a material retranslation. The material shader can still be generated correctly, using the cached translation results, which ensures that material string generation always uses up-to-date input files.

#jira UE-167483
#rb dan.elksnitis, jason.nadro, jason.hoerner, david.doucette, zousar.shaker

[CL 27338134 by massimo tristano in ue5-main branch]