Boolean Query Conditions.
Introduces the ability to combine columns together into a query condition. Example: FColumn<A>() && (FColumn<B>() || FColumn<C>));
At the moment only ANDs and ORs are supported plus scopes. This is enough for the first intended use which is to allow widget constructors to have broader registration conditions with TEDS UI.

#jira UE-194055
#rb brooke.hubert logan.buchy

[CL 27422768 by ronald koppers in ue5-main branch]