Sequencer: add grouping system to help with the common pattern of multiple contributors being grouped when animated the same target
The goal of this change is to replace the ad-hoc grouping and tracking that the property instantiator does with some first-class concept of "grouping", handled by a new system. In turn, this lets the hierarhical bias system look at those generic groups, instead of specifically looking at property blend channels. In the long run, this will let the hierarchical bias system also work with groups of, say, material parameter tracks and material switchers and so on.

#jira UE-187337
#rb andrew.rodham

[CL 27472862 by ludovic chabant in ue5-main branch]