UE-190649 - BP compilation doesn't properly clear loading flags on trashed objects
Not a fix, but adds an ensure that validates the ComponentTemplates array of a Blueprint. The underlying cause of this issue is due to calling PostLoad on a Blueprint before we've resolved placeholder objects. In some cases, the Zen loader can cause this problem, such as during an unexpected LoadPackage call due to resolving a soft pointer.

#jira UE-190649
#rb phillip.kavan,francis.hurteau

[CL 27524813 by dave jones2 in ue5-main branch]