[Backout] - CL27524792 [FYI] dave.jones2
One of the editor tests tripped this unexpectedly. Will check back in once that's resolved.

Original CL Desc
-----------------------------------------------------------------
UE-190649 - BP compilation doesn't properly clear loading flags on trashed objects

Not a fix, but adds an ensure that validates the ComponentTemplates array of a Blueprint. The underlying cause of this issue is due to calling PostLoad on a Blueprint before we've resolved placeholder objects. In some cases, the Zen loader can cause this problem, such as during an unexpected LoadPackage call due to resolving a soft pointer.

#jira UE-190649
#rb phillip.kavan,francis.hurteau

[CL 27533779 by dave jones2 in ue5-main branch]