Sort MessageLog categories alphabetically in the Message Log window
As of writing, the MessageLog categories on the left in the Message Log window are listed in essentially arbitrary order determined some hashmap. This makes it quite difficult for users to navigate the different message log categories and find those of interest to them. One particular workflow that suffers from poor discoverability of messages is reviewing errors and warnings from cloud cooked content when UEFN users submit their data.

At a minimum, this list should be sorted alphabetically to give a predictable order, which is what this CL implements. Future improvements could include bubbling categories with messages to the top or visually marking them in other ways (e.g. color coding).

#nojira
#UEFN-RN Categories in the Message Log window are now sorted alphabetically
#rb monte.krol

[CL 27582450 by sebastian schoner in ue5-main branch]